Use environment inference v2
